+++ R/rtools-metadata.R
@@ -70,6 +70,11 @@ version_info <- list(
   ),
   "4.2" = list(
     version_min = "4.2.0",
+    version_max = "4.2.99",
+    path = "usr/bin"
+  ),
+  "4.3" = list(
+    version_min = "4.3.0",
     version_max = "99.99.99",
     path = "usr/bin"
   ),